refactor: API 설계에 맞춰 거리 double에서 int로 변경

+++ src/main/java/wooteco/subway/domain/Fare.java
@@ -16,7 +16,7 @@ public class Fare {
         this.price = price;
     }
 
-    public static Fare from(final double distance) {
+    public static Fare from(final int distance) {
         return new Fare(calculateFare(distance));
     }
